+++ src/app/core/services/links-item.service.ts
@@ -1,0 +1,75 @@
+/* eslint-disable max-len */
+import { Injectable } from '@angular/core';
+import { AngularFirestore, AngularFirestoreCollection } from '@angular/fire/compat/firestore';
+import { AngularFireAuth } from '@angular/fire/compat/auth';
+
+import { Observable } from 'rxjs';
+
+import { BaseType, IBase } from '@models/base';
+import { ILinkItem, LinkItem } from '@models/link-item';
+import { AppointmentsService } from '@services/appointments.service';
+
+const LINKS_ITEM_COLLECTION = 'enlaces-item';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class LinksItemService {
+
+  private linksItemCollection!: AngularFirestoreCollection<ILinkItem>;
+
+  constructor(
+    private afAuth: AngularFireAuth,
+    private afs: AngularFirestore,
+    private appointmentSrv: AppointmentsService
+  ) {
+    this.linksItemCollection = afs.collection(LINKS_ITEM_COLLECTION);
+  }
+
+  getLinksItemByItemId(itemId: string): Observable<ILinkItem[]> {
+
+    this.linksItemCollection = this.afs.collection<ILinkItem>(
+      LINKS_ITEM_COLLECTION,
+      ref => ref.where('itemId', '==', itemId)
+                .where('active', '==', true)
+                .orderBy('timestamp', 'desc')
+  );
+
+  return this.linksItemCollection.valueChanges();
+}
+
+
+  getLinksItemByRange(dateMin: string, dateMax: string): Observable<ILinkItem[]> {
+
+      this.linksItemCollection = this.afs.collection<ILinkItem>(
+        LINKS_ITEM_COLLECTION,
+        ref => ref.where('active', '==', true)
+                  .where('timestamp', '>=', dateMin)
+                  .where('timestamp', '<=', dateMax)
+                  .orderBy('timestamp', 'desc')
+                  .limit(100)
+    );
+
+    return this.linksItemCollection.valueChanges();
+  }
+
+  // TODO Add Source params
+  async addLinkItem(name: string, item: IBase ): Promise<void> {
+
+      const currentUser = await this.afAuth.currentUser;
+      const linkItem = LinkItem.InitDefault(
+        name,
+        item,
+
+        // TODO: Source interface
+        currentUser.uid,
+        currentUser.displayName,
+        'Usuario registrado'
+        );
+
+      linkItem.id = this.afs.createId();
+      linkItem.timestamp = this.appointmentSrv.getTimestamp();
+      console.log(`linkItem: ${JSON.stringify(linkItem)}`);
+      this.linksItemCollection.doc().set(Object.assign({}, linkItem));
+  }
+}