+++ src/app/features/links/links.module.ts
@@ -1,19 +1,23 @@
 import { NgModule } from '@angular/core';
-import { HttpClientModule } from "@angular/common/http";
+import { HttpClientModule } from '@angular/common/http';
 
 import { SharedModule } from '@shared/shared.module';
 
 import { LinksRoutingModule } from '@features/links/links-routing.module';
 import { LinksComponent } from '@features/links/links.component';
+import { LinksItemComponent } from '@features/links/links-item/links-item.component';
 import { LinkViewComponent } from '@features/links/link-view/link-view.component';
 import { LinkEditComponent } from '@features/links/link-edit/link-edit.component';
 import { LinksDashboardComponent } from '@features/links/links-dashboard/links-dashboard.component';
+import { LinkItemDialogComponent } from '@features/links/link-item-dialog/link-item-dialog.component';
 
 @NgModule({
   declarations: [
     LinksComponent,
     LinkViewComponent,
     LinkEditComponent,
+    LinkItemDialogComponent,
+    LinksItemComponent,
     LinksDashboardComponent,
   ],
   imports: [
@@ -22,6 +26,7 @@ import { LinksDashboardComponent } from '@features/links/links-dashboard/links-d
     LinksRoutingModule,
   ],
   exports: [
+    LinkItemDialogComponent,
     LinksDashboardComponent,
   ]
 })